+++ src/app/products/models/product-model.ts
@@ -3,4 +3,5 @@ export interface ProductModel {
   name: string;
   price: number;
   description: string;
+  available: boolean;
 }