+++ CHANGELOG.md
@@ -6,6 +6,20 @@
 
 ### VZhyrytskiy Вопросы
 
+## [1.0.0] - 2023-05-10
+
+### Added
+
+- Cart Item component
+- WheelEvents in the CartItem
+- Cart model
+- Highlight directive
+- Added modules(Shared, Cart, Products)
+
+
+### Changed
+
+- Changed implementation of the cart
 
 ## [0.0.9] - 2023-04-25
 
@@ -84,7 +98,8 @@
 
 <!-- Versions -->
 
-[unreleased]: https://github.com/m-sobur/Angular36-Sobur/compare/v0.0.9...HEAD
+[unreleased]: https://github.com/m-sobur/Angular36-Sobur/compare/v1.0.0...HEAD
+[0.0.9]: https://github.com/m-sobur/Angular36-Sobur/compare/v0.0.9...v1.0.0
 [0.0.9]: https://github.com/m-sobur/Angular36-Sobur/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/m-sobur/Angular36-Sobur/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/m-sobur/Angular36-Sobur/compare/v0.0.6...v0.0.7